Filter out events that close the auto-completion widget
If the user presses a key and the auto-completion widget disappears somehow,
the user doesn't want that the client application also performs an action that he
or she does not know about.

For instance, in Dolphin, pressing Enter closed the completion widget and also
triggered a search. Now, the completion widget is closed and no unwanted search
is performed.