Simpler sizing of the completion widget, and fix an infinite recursion
If UniformItemSizes is set to false, opening the completion widget requires
all the items to be laid out again. Something in this operation causes an
infinite recursion of signals sent between the delegate items and the list view
itself, and the program crashes. Now, items are laid out one time and never touched
after that, and everything works.